+++ OneWoW/Locales/enUS.lua
@@ -248,11 +248,6 @@ ns.Locale:Register(ADDON_NAME, "enUS", {
     ["DATA_MOD_PROFESSIONS"] = "Professions",
     ["DATA_MOD_STORAGE"] = "Storage",
 
-    ["CAT_MOD_JOURNAL"] = "Journal",
-    ["CAT_MOD_TRADESKILLS"] = "Tradeskills",
-    ["CAT_MOD_QUESTS"] = "Quests",
-    ["CAT_MOD_QUESTS_ARCHIVE"] = "Quest Archive",
-    ["CAT_MOD_VENDORS"] = "Vendors",
     ["CAT_MOD_ZONEDB"] = "Zone Database",
     ["CAT_MOD_NPCDB"] = "NPC Database",
     ["CAT_MOD_ITEMDB"] = "Item Database",