+++ OneWoW/Locales/zhTW.lua
@@ -248,11 +248,6 @@ ns.Locale:Register(ADDON_NAME, "zhTW", {
     ["DATA_MOD_PROFESSIONS"] = "Professions",
     ["DATA_MOD_STORAGE"] = "Storage",
 
-    ["CAT_MOD_JOURNAL"] = "Journal",
-    ["CAT_MOD_TRADESKILLS"] = "Tradeskills",
-    ["CAT_MOD_QUESTS"] = "Quests",
-    ["CAT_MOD_QUESTS_ARCHIVE"] = "任務封存",
-    ["CAT_MOD_VENDORS"] = "Vendors",
     ["CAT_MOD_ZONEDB"] = "區域資料庫",
     ["CAT_MOD_NPCDB"] = "NPC資料庫",
     ["CAT_MOD_ITEMDB"] = "物品資料庫",